requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
auth0-python==4.8.1
azure-core==1.32.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb
click==8.1.8
cohere
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.108.0
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==3.2.0
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
email_validator==2.2.0
embedchain==0.1.128
et_xmlfile==2.0.0
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.6
fastavro==1.10.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.9.0
gptcache==0.1.44
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
instructor==1.7.7
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.10.0
json_repair==0.40.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
lancedb==0.21.2
langchain==0.3.20
langchain-anthropic==0.3.10
langchain-cohere==0.3.5
langchain-community==0.3.19
langchain-core==0.3.49
langchain-experimental==0.3.4
langchain-google-genai==2.1.0
langchain-openai==0.2.14
langchain-text-splitters==0.3.6
langgraph==0.3.5
langsmith==0.3.19
litellm
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.65.5
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.0.0
portalocker==2.10.1
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.13.3
# Editable install with no version control (react-agent==0.0.1)
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
schema==0.7.7
scipy==1.15.2
sentence-transformers==3.3.1
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
timm==1.0.12
tokenizers
tomli==2.2.1
tomli_w==1.2.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.10
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0

PyGithub~=2.6.1
defusedxml~=0.7.1
trio~=0.29.0
outcome~=1.3.0.post0
selenium~=4.30.0
webdriver-manager~=4.0.2
PySocks~=1.7.1
pyOpenSSL~=25.0.0
hyperlink~=21.0.0